Added the Task class and added it to ChildManager
Added the Task class with fields for the name of the task (String), a list of UUIDs to keep track of who is next for the task, and a String to easily get the name of the child who is next. Added methods to cycle UUIDs upon completion, remove a child that has been deleted, set the next child, and setters as needed. Added a list of tasks to ChildManager and implemented methods to get needed data and to delete children from tasks when they are deleted in allChildren